+++ src/AppRouter.jsx
@@ -16,6 +16,7 @@ import Logout from './logout.jsx'
 import UserManagement from './userManagement.jsx'
 import PPPPage from './ppp.jsx'
 import CameraSwitcherPage from './cameraswitcher.jsx'
+import PipelineEditorPage from './pipelineeditor.jsx'
 
 function AppRouter () {
   const [isAuthenticated, setIsAuthenticated] = useState(null)
@@ -76,6 +77,7 @@ function AppRouter () {
           <Link className='list-group-item list-group-item-action bg-light' to="/apclients">Access Point Clients</Link>
           <Link className='list-group-item list-group-item-action bg-light' to="/video">Photo and Video</Link>
           <Link className='list-group-item list-group-item-action bg-light' to="/cameraswitcher">Camera Switcher</Link>
+          <Link className='list-group-item list-group-item-action bg-light' to="/pipelineeditor">Video Pipeline Editor</Link>
           <Link className='list-group-item list-group-item-action bg-light' to="/cloud">Cloud Upload</Link>
           <Link className='list-group-item list-group-item-action bg-light' to="/vpn">VPN Config</Link>
           <Link className='list-group-item list-group-item-action bg-light' to="/about">About</Link>
@@ -96,6 +98,7 @@ function AppRouter () {
             <Route exact path="/about" element={<About />} />
             <Route exact path="/video" element={<VideoPage />} />
             <Route exact path="/cameraswitcher" element={<CameraSwitcherPage />} />
+            <Route exact path="/pipelineeditor" element={<PipelineEditorPage />} />
             <Route exact path="/flightlogs" element={<LogBrowser />} />
             <Route exact path="/apclients" element={<NetworkClients />} />
             <Route exact path="/ntrip" element={<NTRIPPage />} />